requirements.txt
Django==5.1.3
Djangorestframework==3.15.2
black==24.10.0
flake8==7.1.1